+++ src/main/java/com/yoyomo/infra/redis/RetryQueueProcessor.java
@@ -2,6 +2,7 @@ package com.yoyomo.infra.redis;
 
 import com.yoyomo.domain.club.exception.UnavailableSubdomainException;
 import com.yoyomo.infra.aws.s3.service.S3Service;
+import com.yoyomo.infra.redis.service.RedisQueueService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
