+++ src/main/java/com/yoyomo/infra/redis/service/RedisQueueService.java
@@ -1,4 +1,4 @@
-package com.yoyomo.infra.redis;
+package com.yoyomo.infra.redis.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
